[2021-11-25] ADD  image_board

+++ jspstudy/workspace/11_MYHOME/src/dto/Free.java
@@ -13,7 +13,7 @@ public class Free {
 	private int state;
 	private int depth;
 	private Long groupNo;
-	private int groupOrd;
+	private Long groupOrd;
 	
 	
 	
@@ -77,10 +77,11 @@ public class Free {
 	public void setGroupNo(Long groupNo) {
 		this.groupNo = groupNo;
 	}
-	public int getGroupOrd() {
+	
+	public Long getGroupOrd() {
 		return groupOrd;
 	}
-	public void setGroupOrd(int groupOrd) {
+	public void setGroupOrd(Long groupOrd) {
 		this.groupOrd = groupOrd;
 	}
